+++ src/profiles.ts
@@ -6,7 +6,7 @@ import { StreamDeckConnection } from "./connectivity/connection";
 export class ProfileClient {
 	/**
 	 * Initializes a new instance of the {@link ProfileClient} class.
-	 * @param connection Underlying connection with the Stream Deck.
+	 * @param connection Connection with Stream Deck.
 	 */
 	constructor(private readonly connection: StreamDeckConnection) {}
